@@ -1,11 +1,11 @@
 name: midtransflutter
 description: Midtrans Payment Gateway for Flutter (Only Support 3DS Payment)
-version: 0.0.1
+version: 1.0.0
 author:
 homepage:
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
